难道就我遇到了这个问题?
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@11061 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/php/phpbbs_article.c
@@ -355,7 +355,7 @@ PHP_FUNCTION(bbs_get_threads_from_gid)
     long start;
     zval *z_threads;
     zval *retprev;
-    int ins_top = 0;
+    long ins_top = 0;
     int i, j;
     const struct boardheader *bp;
     int is_bm;
